refactoring the file structure
