+++ browser/game/controls/gameMaster.js
