+++ browser/game/engine/create.js
@@ -1,37 +1,17 @@
-import store from '../store'
-import Survivor from './survivorMode.js';
-import GameMaster from './gameMasterMode.js';
+import store from '../../store'
+import Survivor from '../controls/survivor.js';
+import GameMaster from '../controls/gameMaster.js';
+import socket from '../../socket'
+import Bullets from '../entities/bullets.js';
 
 let player, bullets, playerCollide, teamBullet, survivor, gameMaster;
-let teammates = {} //TODO: on the backend .on('connection'), populate this with existing players instead of waiting for the first interval
-
-// import Player from './entities/players.js';
-import Bullets from './entities/bullets.js';
-// // import Wall from './entities/mapObjects.js';
-// import socket from '../socket';
-
-
-// var player;
-// var walls;
-// var cursors;
-// var wasd;
-// var fireRate = 400;
-// var monsterRate = 1000;
-// var button;
-// var bullets;
-// let id = 0;
-
-// let teammates = {};
-// let map, groundLayer, featuresBottom, playerOnBottom, playerOnTop, playerBehindBottom, playerBehindTop, playerCollide;
-//TODO: on the backend .on('connection'), populate this with existing players instead of waiting for the first interval
-//120*32
-//80*32
+let teammates = {}
 
 export default function create() {
     //this settings
     this.world.setBounds(-1920, -1280, 3840, 2560);
     this.physics.startSystem(Phaser.Physics.ARCADE);
-    
+
     // map, order matters!
     let map = this.add.tilemap('tilemap');
     map.addTilesetImage('terrain_atlas', 'tileset');
@@ -62,6 +42,14 @@ export default function create() {
     teamBullet = new Bullets(this);
     //button
     //button = this.add.button(this.world.centerX - 95, 400, 'button', spawn, this, 2, 1, 0);
+
+    const emitClient = () => {
+      setInterval(() => {
+          let state = store.getState();
+          socket.emit('send_all_data', state);
+        }, 1000/30);
+    }
+    emitClient()
 }
 
 export {player, bullets, teammates, teamBullet, playerCollide, survivor, gameMaster};
