+++ browser/index.js
@@ -1,13 +1,9 @@
-//TODO: make this the entry point for the frontend,
-//rendering app and invoking sockets
 import 'pixi';
 import 'p2';
 import 'phaser';
+import { initializeSocket } from './socket';
 
-import game from './game/game.js';
-import store from './store';
-import { initializeSocket, emitClient } from './socket';
+//start the game!
+import game from './game/engine/StartGame.js';
 
 initializeSocket();
-emitClient();
-
